+++ rest/src/main/java/pl/piomin/services/rest/customer/service/ImageService.java
@@ -30,15 +30,18 @@ public class ImageService {
 
     public byte[] downloadImage(String imageName) {
         Optional<Image> dbImage = imageRepository.findByName(imageName);
-        imageRepository.deleteById(dbImage.get().getId());
-        return dbImage.map(image -> {
-            try {
-                return ImageUtils.decompressImage(image.getImageData());
-            } catch (DataFormatException | IOException exception) {
-                throw new ContextedRuntimeException("Error downloading an image", exception)
-                        .addContextValue("Image ID",  image.getId())
-                        .addContextValue("Image name", imageName);
-            }
-        }).orElse(null);
+        if(dbImage.isPresent()){
+            imageRepository.deleteById(dbImage.get().getId());
+            return dbImage.map(image -> {
+                try {
+                    return ImageUtils.decompressImage(image.getImageData());
+                } catch (DataFormatException | IOException exception) {
+                    throw new ContextedRuntimeException("Error downloading an image", exception)
+                            .addContextValue("Image ID",  image.getId())
+                            .addContextValue("Image name", imageName);
+                }
+            }).orElse(null);
+        }
+        return null;
     }
 }
